comptes services

+++ src/main/java/uphf/banque/entities/rest/compte/PostPaiementRequest.java
@@ -8,7 +8,7 @@ import lombok.Setter;
 @Setter
 @Builder
 public class PostPaiementRequest {
-    private double montant;
+    private float montant;
 
     private String dateCreation;
 }
